requirements.txt
flask
alpaca-trade-api
python-dotenv
gunicorn